simple test passes
